requirements.txt
openai>=1.0.0
PyPDF2>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0
opencv-contrib-python>=4.5.0
pyzbar>=0.1.9
streamlit>=1.39.0
Pillow>=10.0.0
